Deploys project on GKE 1.15.7
This change fixes a problem whereby deployment of the marketplace
app would fail on GKE clusters with Kubernetes versions >= 1.15.

The main problem was that we were using the Google Marketplace Tools
Version 0.7.0, and this uses a version of kubectl (1.12) that does
not know how to authenticate with Kubernetes version 1.15 or newer.

The changes included are as follows:

* Upgraded the Google Marketplace Tools container image used by the Deployer from 0.7.0 to 0.10.0. The newer version of these tools provide an adaptive Kubectl client version (tools read the Kubernetes server version, then select a matching kubectl binary).
* Upgraded the Google Marketplace Tools submodule to 0.10.0.
* Upgraded the version of Helm used by the deployer from 2.6.1 to 2.16.1 to eliminate this Helm bug: https://github.com/helm/helm/issues/2998
* Deleted x-google-marketplace section for tester.image in schema.yaml to be consistent for Google Marketplace Tools v0.10.0
* Added a build.sh flag (`-p` or `--persist`) to persist the application deployment after testing.
* Fixed the deployment's `deploy-info` annotations to use keys that are valid JSON keys (i.e. with quotes).

Fixes Issue #25 "Project fails to deploy on GKE 1.15.7-gke.23"