Adds trivy scan of component images in Jenkins pipeline
Adds a trivy security scan of the following Marketplace app component images
to project pipeline to find any Vulnerabilities before publishing to
Google Marketplace:
- deployer
- tester
- nginx
- postgres

A scan of the conjur image is skipped because this image is built and
scanned elsewhere (i.e. for the cyberark/conjur repo builds).

A .trivyignore file is being added for the following remaining
critical vulnerabilities:
- 4 critical vulnerability remaining in Google's deployer_helm container
- 7 critical vulnerabilities in postgres:10.1 container
- 3 critical vulnerabilities in nginx:1.15 container

NOTE: If we make the following image changes:
- Downgrade to postgres:9.4 (which would be consistent with Conjur OSS)
- Upgrace to nginx:1.17
then the critical vulnerabilities will be reduced to 2 for both
postgres and nginx images.

Reference: https://github.com/aquasecurity/trivy