Reduces vulnerabilities by updating deployer_helm container
This change updates the version of the Google `deployer_helm` container
used as a baseline container in our Google Marketplace app deployer
container from version 0.10.0 to 0.10.1.

This is being done to reduce the number of critical and high
vulnerabilities that are being reported when a trivy scan is run on our
deployer container, as described in Issue #34.

With this change:
- Critical vulnerabilities are reduced from 3 to 1 (unique CVEs)
- High vulnerabilities are reduced from 17 to 14 (unique CVEs)