Conjur version is rolled back to 1.5.0
This change rolls back the release version of Conjur open source that
is included in the Marketplace application from 1.6.0 back to 1.5.0.
This is being done because the release 1.6.0 is not deemed to be stable.

Some other changes:

* Marketplace application version is changed to 1.5.0 to match Conjur
release being used

* As images are built, they are pushed to a registry location that uses
`cyberark` as a repo name, i.e.:

```
https://gcr.io/conjur-cloud-launcher-onboard/cyberark:1.6.1
https://gcr.io/conjur-cloud-launcher-onboard/cyberark/deployer:1.6.1
https://gcr.io/conjur-cloud-launcher-onboard/cyberark/nginx:1.6.1
https://gcr.io/conjur-cloud-launcher-onboard/cyberark/postgres:1.6.1
https://gcr.io/conjur-cloud-launcher-onboard/cyberark/tester:1.6.1
```

This is a bit confusing, but the way our Google Marketplace app was
initially set up, the Google Marketplace publishing tools look for
images to publish in the locations shown above. I checked with
Google Marketplace engineers on this, and they say that the above
locations are correct, and are as they expect.

* In CONTRIBUTING.md, the step to check for vulnerabilities on the
GCR is deleted, since Google no longer checks for vulnerabilities
in Marketplace images by default. (This is a paid service now.)

* The Google marketplace helm deployer that serves as a baseline for
our deployer has been downgraded to 0.9.10, since version 0.10.0 and
0.10.1 have vulnerabilities that we want to avoid.

Fixes Issue #50