Jenkins fails when smoketests in test pods fail
This change fixes a problem whereby the Jenkins CI job was not being
marked as a failure when the smoketests in the test pods fail.

This appears to be a bug in the Google marketplace tools. The script
that's being used to deploy and test Conjur is
`marketplace-k8s-app-tools/scripts/verify`. The problem with this script
is that it checks whether the `deployer` job succeeds, but it doesn't
check for success/failure of the two test pods that are created.

Fixes Issue #49